% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtable.R
\name{runtable}
\alias{runtable}
\title{Load a dataframe  + list options: general_list +  table_list_css_header + list2 not implemented +  table_list_css_column + table_list_css_column_single

This function loads a dataframe and option list}
\usage{
runtable(data, opts = NULL, ...)
}
\arguments{
\item{data}{Dataframe}

\item{opts}{List of list, see the structure examples (nested)}
}
\value{
Datatable
}
\description{
Load a dataframe  + list options: general_list +  table_list_css_header + list2 not implemented +  table_list_css_column + table_list_css_column_single

This function loads a dataframe and option list
}
