% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabm-package.R
\docType{package}
\name{tabm-package}
\alias{tabm}
\alias{tabm-package}
\title{tabm: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Felipe Carranza \email{fcarranza@protonmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
